pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rechunk"
version = "0.1.0"
description = "Feedback-driven RAG chunking: adapt chunk strategies without full reindex; LlamaIndex + Temporal"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ReChunk" }]
keywords = ["rag", "chunking", "llamaindex", "retrieval"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.2.0",
    "llama-index-embeddings-openai>=0.2.0",
    "PyPDF2",
    "python-docx",
    "temporalio",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
# Hugging Face export for benchmark corpora (scripts/prepare_hf_benchmark_corpus.py)
# datasets 4.x removed loading-script datasets (e.g. deepmind/pg19). Keep <4 until those hubs ship Parquet-only layouts.
benchmark-corpora = [
    "datasets>=2.16.0,<4.0.0",
    # CUAD on Hugging Face includes Pdf-typed columns; datasets may decode them while iterating.
    "pdfplumber>=0.11.0",
]

[project.urls]
Repository = "https://github.com/rechunk/rechunk"
Documentation = "https://github.com/rechunk/rechunk#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
